versao 0.9

+++ src/Forms/Forms_principais/Form_configuracoes/sub_area/consultar_subarea.cs
@@ -17,41 +17,16 @@ namespace PSI18H_M16_Projeto_2218088_RodrigoBarata.Forms.Forms_principais.Form_c
         public consultar_subarea()
         {
             InitializeComponent();
-            combobox();
+            
             combobox_consultar();
             dataview();
         }
         DB db = new DB();
-        void combobox()
-        {
-            string selectQuery = "SELECT * FROM area ORDER BY nome_area ASC;";
-            using (MySqlCommand mysqlcommand = new MySqlCommand(selectQuery, db.connection))
-            {
-                MySqlDataReader myReader;
-                try
-                {
-                    db.openConnection();
-                    myReader = mysqlcommand.ExecuteReader();
-                    DataTable dt = new DataTable();
-                    dt.Load(myReader);
-                    cbxarea.DisplayMember = "nome_area";
-                    cbxarea.ValueMember = "idarea";
-                    cbxarea.DataSource = dt;
-                }
-                catch (Exception erro)
-                {
-                    MessageBox.Show("Erro:" + erro.Message);
-                }
-                finally
-                {
-                    db.closeConnection();
-                }
-            }
-        }
+        
 
         public void dataview()
         {
-            string selectQuery = "SELECT* FROM subarea WHERE area_idarea = " + cbxareas.SelectedValue;
+            string selectQuery = "SELECT idsubarea, nome_subarea FROM subarea WHERE area_idarea = " + cbxareas.SelectedValue;
             DataTable table = new DataTable();
             MySqlDataAdapter adapter = new MySqlDataAdapter(selectQuery, db.connection);
             adapter.Fill(table);
@@ -88,57 +63,126 @@ namespace PSI18H_M16_Projeto_2218088_RodrigoBarata.Forms.Forms_principais.Form_c
             }
         }
 
+        public Boolean checkTextBoxesValues()
+        {
+            if (txtsubarea.Text.Equals("") || cbxareas.SelectedValue.Equals(""))
+            {
+                return false;
+            }
+            else
+            {
+                return true;
+            }
+        }
+
+        public Boolean checkSubArea()
+        {
+            DB db = new DB();
+
+            String nome = txtsubarea.Text;
+
+
+            DataTable table = new DataTable();
+
+            MySqlDataAdapter adapter = new MySqlDataAdapter();
+
+            MySqlCommand command = new MySqlCommand("SELECT * FROM subarea WHERE nome_subarea = @nome", db.getConnection());
+
+            command.Parameters.Add("@nome", MySqlDbType.VarChar).Value = nome;
+
+            adapter.SelectCommand = command;
+
+            adapter.Fill(table);
+            if (table.Rows.Count > 0)
+            {
+                return true;
+            }
+            else
+            {
+                return false;
+            }
+        }
+
         private void btnnovo_Click(object sender, EventArgs e)
         {
             string insertQuery = "INSERT INTO subarea ( nome_subarea, area_idarea) VALUES ( @param1, @param2)";
             using (MySqlCommand cmd = new MySqlCommand(insertQuery, db.connection))
             {
-               
-                try
-                {
-                    db.openConnection();
 
-                    cmd.Parameters.AddWithValue("@param1", txtsubarea.Text);
-                    cmd.Parameters.AddWithValue("@param2", cbxarea.SelectedValue);
-                    cmd.ExecuteNonQuery();
-                    dataview();
-                    db.closeConnection();
-                }
-                catch (Exception erro)
+                if (checkTextBoxesValues())
                 {
-                    MessageBox.Show("Erro:" + erro.Message);
+                    if (checkSubArea())
+                    {
+                        MessageBox.Show("Já existe esta Subárea, escolha outra");
+                    }
+                    else
+                    {
+                        try
+                        {
+                            db.openConnection();
+
+                            cmd.Parameters.AddWithValue("@param1", txtsubarea.Text);
+                            cmd.Parameters.AddWithValue("@param2", cbxareas.SelectedValue);
+                            cmd.ExecuteNonQuery();
+                            dataview();
+                            db.closeConnection();
+                        }
+                        catch (Exception erro)
+                        {
+                            MessageBox.Show("Erro:" + erro.Message);
+                        }
+                        finally
+                        {
+                            db.closeConnection();
+                        }
+                    }
                 }
-                finally
+                else
                 {
-                    db.closeConnection();
+                    MessageBox.Show("Preencha os campos");
                 }
+                
             }
 
         }
 
         private void btneditar_Click(object sender, EventArgs e)
         {
-            string updateQuery = "UPDATE `subarea` SET `nome_subarea`='"+txtsubarea.Text+"',`area_idarea`='"+cbxarea.SelectedValue+ "' WHERE idsubarea =" + int.Parse(txtid.Text);
+            string updateQuery = "UPDATE `subarea` SET `nome_subarea`='"+txtsubarea.Text+"',`area_idarea`='"+cbxareas.SelectedValue+ "' WHERE idsubarea =" + int.Parse(txtid.Text);
             using (MySqlCommand cmd = new MySqlCommand(updateQuery, db.connection))
             {
+                if (checkTextBoxesValues())
+                {
+                    if (checkSubArea())
+                    {
+                        MessageBox.Show("Já existe esta Subárea, escolha outra");
+                    }
+                    else
+                    {
+                        try
+                        {
+                            db.openConnection();
 
-                try
-                {
-                    db.openConnection();
 
-                    
-                    cmd.ExecuteNonQuery();
-                    dataview();
-                    db.closeConnection();
+                            cmd.ExecuteNonQuery();
+                            dataview();
+                            db.closeConnection();
+                        }
+                        catch (Exception erro)
+                        {
+                            MessageBox.Show("Erro:" + erro.Message);
+                        }
+                        finally
+                        {
+                            db.closeConnection();
+                        }
+                    }
                 }
-                catch (Exception erro)
+                else
                 {
-                    MessageBox.Show("Erro:" + erro.Message);
+                    MessageBox.Show("Preencha os campos");
                 }
-                finally
-                {
-                    db.closeConnection();
-                }
+               
             }
         }
 
@@ -150,35 +194,50 @@ namespace PSI18H_M16_Projeto_2218088_RodrigoBarata.Forms.Forms_principais.Form_c
 
         private void btnremover_Click(object sender, EventArgs e)
         {
-            string deleteQuery = "DELETE FROM subarea WHERE idsubarea = " + int.Parse(txtid.Text);
-            using (MySqlCommand cmd = new MySqlCommand(deleteQuery, db.connection))
-            {
-
-                try
+           
+                if ( checkTextBoxesValues())
                 {
-                    db.openConnection();
+                if (checkSubArea())
+                {
+                    MessageBox.Show("Esta Suárea contém cursos\n Elimine primeiro os Cursos");
+                }
+                else
+                {
+                    try
+                    {
+                        string deleteQuery = "DELETE FROM subarea WHERE idsubarea = " + int.Parse(txtid.Text);
+                        using (MySqlCommand cmd = new MySqlCommand(deleteQuery, db.connection))
+                        {
+                            db.openConnection();
 
 
-                    cmd.ExecuteNonQuery();
-                    dataview();
-                    db.closeConnection();
+                            cmd.ExecuteNonQuery();
+                            dataview();
+                            db.closeConnection();
+                        }
+                    }
+                    catch (Exception erro)
+                    {
+                        MessageBox.Show("Erro:" + erro.Message);
+                    }
+                    finally
+                    {
+                        db.closeConnection();
+                    }
                 }
-                catch (Exception erro)
+                }
+                else
                 {
-                    MessageBox.Show("Erro:" + erro.Message);
+                    MessageBox.Show("Preencha os campos");
                 }
-                finally
-                {
-                    db.closeConnection();
-                }
-            }
+            
         }
 
         private void btnsearch_Click(object sender, EventArgs e)
         {
             MySqlDataReader mdr;
             
-            string select = "SELECT * FROM area WHERE area = " + txtconsultararea.Text;
+            string select = "SELECT * FROM subarea WHERE nome_subarea like '%" + txtconsultararea.Text +"%'";
             command = new MySqlCommand(select, db.connection);
             db.openConnection();
             mdr = command.ExecuteReader();
@@ -186,13 +245,18 @@ namespace PSI18H_M16_Projeto_2218088_RodrigoBarata.Forms.Forms_principais.Form_c
             if (mdr.Read())
             {
 
-                txtsubarea.Text = mdr.GetString("nome_area");
+                txtsubarea.Text = mdr.GetString("nome_subarea");
             }
             else
             {
-                MessageBox.Show("Área não encontrada");
+                MessageBox.Show("Subárea não encontrada");
             }
             db.closeConnection();
+        }
+
+        private void cbxareas_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            dataview();
         }
     }
 }
